Split Fifteen and Pennies demos into separate files.
