+++ jorendorff+haskell/.gitignore
@@ -2,7 +2,7 @@
 *.o
 *.hi
 Reversi
-MinimaxDemo
 Chess
-
-
+Pennies
+Fifteen
+ChessTest
